// src/app/page/model/xmlns/content/content-ct-text.ts
export interface KgwContentComplexTypeText {
  value?: string;
  attributes?: {
    startImage?: string;
    startImageSize?: number;
    endImage?: string;
    endImageSize?: number;
    i18n_id?: string;
    text_align?: string;
    text_color?: string;
    text_scale?: string;
    text_style?: string;
    version?: number;
    restrictTo?: Array<string>;
  };
}